Refactor DWG object writer: centralize and validate enqueue
Moved private writer methods from DwgObjectWriter.Common.cs to DwgObjectWriter.cs for better organization. Introduced enqueueValidObject to ensure only valid NonGraphicalObject instances are enqueued, replacing direct queue usage. Updated XDictionary handling and removed the obsolete _dictionaries field. Improved encapsulation and robustness of the DWG object writing process.
